requirements.txt
flask
schedule
requests
gspread
oauth2client
python-dotenv
pytz
